refactors the mobile nav dom structure
