+++ components/src/NavBar/NavBar.js
@@ -3,6 +3,7 @@ import PropTypes from "prop-types";
 import styled from "styled-components";
 import { Box } from "../Box";
 import { Flex } from "../Flex";
+import { Icon } from "../Icon";
 import NavBarSearch from "../NavBarSearch/NavBarSearch";
 import Branding from "./Branding";
 import DesktopMenu from "./DesktopMenu";
@@ -10,11 +11,7 @@ import MobileMenu from "./MobileMenu";
 import { withMenuState } from "./withMenuState";
 import isValidMenuItem from "./isValidMenuItem";
 import theme from "../theme";
-
-const navBarStyles = {
-  background: theme.colors.blackBlue,
-  padding: `${theme.space.x2} ${theme.space.x3}`,
-};
+import { subPx } from "../Utils"
 
 const MediumNavBar = ({
   menuData,
@@ -24,23 +21,21 @@ const MediumNavBar = ({
 }) => (
   <Box { ...props }>
     <Branding desktopSrc={ desktopSrc } alt={ alt } />
-    <nav>
-      <Flex justifyContent="space-between" alignContent="flex-end">
-        {menuData.primaryMenu
-        && (
-        <Flex alignItems="center" pr="x3">
-          <DesktopMenu menuData={ menuData.primaryMenu } />
-        </Flex>
-        )}
-        <Box width={ 1 }>
-          <Flex style={ { "float": "right" } }>
-            {menuData.search && <Flex maxWidth="18em"><NavBarSearch { ...menuData.search } /></Flex>}
-            {menuData.secondaryMenu
-            && <DesktopMenu pl="x2" menuData={ menuData.secondaryMenu } />}
-          </Flex>
-        </Box>
+    <Flex justifyContent="space-between" alignContent="flex-end" style={ { flexGrow: "1", margin: `0 0 0 ${theme.space.x3}` } }>
+      {menuData.primaryMenu
+      && (
+      <Flex alignItems="center" pr="x3">
+        <DesktopMenu menuData={ menuData.primaryMenu } />
       </Flex>
-    </nav>
+      )}
+      <Box width={ 1 }>
+        <Flex style={ { "float": "right" } }>
+          {menuData.search && <Flex maxWidth="18em"><NavBarSearch { ...menuData.search } /></Flex>}
+          {menuData.secondaryMenu
+          && <DesktopMenu pl="x2" menuData={ menuData.secondaryMenu } />}
+        </Flex>
+      </Box>
+    </Flex>
   </Box>
 );
 
@@ -56,27 +51,62 @@ MediumNavBar.defaultProps = {
   menuData: null,
 };
 
+const MobileMenuTrigger = styled.button(
+  {
+    color: theme.colors.white,
+    background: "none",
+    border: "none",
+    padding: `${subPx(theme.space.x1)} ${theme.space.x1}`,
+    marginLeft: theme.space.x1,
+    borderRadius: theme.radii.medium,
+    transition: ".2s",
+    "&:hover, &:focus": {
+      outline: "none",
+      color: theme.colors.lightBlue,
+      backgroundColor: theme.colors.black,
+      cursor: "pointer",
+    },
+  }
+);
+
 const SmallNavBar = withMenuState(({
   menuData,
-  menuState,
+  menuState: { isOpen, handleMenuToggle, closeMenu },  
   mobileSrc,
   alt,
   style,
   ...props
 }) => (
-  <Box { ...props } style={ Object.assign({}, { height: menuState.isOpen ? "100vh" : null, position: "relative", overflow: "auto" }, style) }>
+  <>
+  <Box { ...props } style={ Object.assign({}, { position: "relative", overflow: "auto" }, style) }>
     <Branding mobileSrc={ mobileSrc } alt={ alt } />
-    <nav>
-      <Flex justifyContent="flex-end">
-        {menuData.search && <Flex maxWidth="18em" alignItems="center" px="0"><NavBarSearch { ...menuData.search } /></Flex>}
-        {
-          (menuData.primaryMenu || menuData.secondaryMenu)
-            && <MobileMenu menuData={ menuData } menuState={ menuState } display="block" />
-        }
-      </Flex>
-    </nav>
+    <Flex justifyContent="flex-end" style={ { flexGrow: "1", margin: `0 0 0 ${theme.space.x3}` } }>
+      {menuData.search && 
+        <Flex maxWidth="18em" alignItems="center" px="0">
+          <NavBarSearch { ...menuData.search } />
+        </Flex>
+      }
+      {(menuData.primaryMenu || menuData.secondaryMenu) && 
+        <MobileMenuTrigger onClick={ handleMenuToggle } aria-expanded={ isOpen ? true : null }>
+          {
+          isOpen
+            ? <Icon icon="close" title="Close Menu" />
+            : <Icon icon="menu" title="Open Menu" />
+          }
+        </MobileMenuTrigger>
+      }
+    </Flex>
   </Box>
+      {(isOpen) && 
+        <MobileMenu menuData={ menuData } closeMenu={ closeMenu }/>
+      }
+    </>
 ));
+
+const navBarStyles = {
+  background: theme.colors.blackBlue,
+  padding: `${theme.space.x2} ${theme.space.x3}`,
+};
 
 const BaseNavBar = ({
   menuData,
@@ -104,11 +134,6 @@ BaseNavBar.defaultProps = {
   className: null,
 };
 
-const NavBar = styled(BaseNavBar)({
-  "nav": {
-    flexGrow: "1",
-    margin: `0 0 0 ${theme.space.x3}`,
-  },
-});
+const NavBar = styled(BaseNavBar)({});
 
 export default NavBar;
